reg_service/requirements.txt
Flask
psycopg2-binary
kubernetes
loguru